requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
bidict==0.23.1
bitsandbytes==0.42.0
blinker==1.9.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cudf-cu12==25.6.0
cuml-cu12==25.6.0
cupy-cuda12x==13.5.1
cuvs-cu12==25.6.1
cycler==0.12.1
dask==2025.5.0
dask-cuda==25.6.0
dask-cudf-cu12==25.6.0
decorator==4.4.2
decord==0.6.0
deepspeed==0.17.3
distributed==2025.5.0
distributed-ucxx-cu12==0.44.0
draccus==0.8.0
ema-pytorch==0.5.1
etils==1.13.0
evdev==1.9.2
fastrlock==0.8.3
filelock==3.18.0
Flask==3.1.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.9.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.5
hjson==3.1.0
huggingface-hub==0.26.1
hydra-core==1.3.2
idna==3.10
imageio==2.34.2
imageio-ffmpeg==0.6.0
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonlines==4.0.0
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
libcudf-cu12==25.6.0
libcuml-cu12==25.6.0
libcuvs-cu12==25.6.1
libkvikio-cu12==25.6.0
libraft-cu12==25.6.0
librmm-cu12==25.6.0
libucx-cu12==1.18.1
libucxx-cu12==0.44.0
lightning==2.4.0
lightning-utilities==0.15.0
llvmlite==0.44.0
locket==1.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mergedeep==1.3.4
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
mujoco==3.3.4
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numba-cuda==0.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.12
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.11.1
packaging==24.1
pandas==2.2.3
partd==1.4.2
peft==0.10.0
pillow==11.2.1
piq==0.8.0
platformdirs==4.3.8
portalocker==3.2.0
proglog==0.1.12
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pylibcudf-cu12==25.6.0
pylibraft-cu12==25.6.0
pynput==1.8.1
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
PyOpenGL==3.1.9
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
python-xlib==0.33
pytorch-lightning==1.8.6
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
raft-dask-cu12==25.6.0
rapids-dask-dependency==25.6.0
rapids-logger==0.1.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rmm-cu12==25.6.0
robosuite==1.4.1
rotary-embedding-torch==0.8.4
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.1.99
sentry-sdk==2.34.1
shapely==2.1.1
shortuuid==1.0.13
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tblib==3.1.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.9.12
tokenizers==0.15.1
toml==0.10.2
toolz==1.0.0
torch==2.2.0
torchmetrics==0.10.3
torchvision==0.17.0
tornado==6.5.1
tqdm==4.66.4
transformers==4.37.2
treelite==4.4.1
triton==2.2.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ucx-py-cu12==0.44.0
ucxx-cu12==0.44.0
urllib3==2.5.0
wandb==0.21.0
webdataset==0.2.111
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.24
yacs==0.1.8
yarl==1.20.1
zict==3.0.0
zipp==3.23.0
wandb==0.21.0
einops-exts==0.0.4
imgaug==0.4.0
tensorboard==2.15.2
-e /work/Depth-Anything/torchhub/facebookresearch_dinov2_main
